examples/hubspot-crm/requirements.txt
mnemora-sdk>=0.1.2
httpx>=0.27
anthropic>=0.40
python-dotenv>=1.0
rich>=13.0
